RegisterFrom refractor initial setup
